# docs/swagger.yaml
basePath: /
definitions:
  fiber.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  go-users-api_internal_models.User:
    properties:
      age:
        type: integer
      dob:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  internal_handler.CreateUserRequest:
    properties:
      dob:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - dob
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing users and calculating age dynamically
  title: Go Users API
  version: "1.0"
paths:
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-users-api_internal_models.User'
            type: array
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a user with name and date of birth
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-users-api_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Delete user
      tags:
      - Users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-users-api_internal_models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-users-api_internal_models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Update user
      tags:
      - Users
swagger: "2.0"